Extract modal types, utils
