+++ src/__tests__/components/CreateProjectModal.test.tsx
@@ -8,26 +8,6 @@ import papi from '@papi/frontend';
 import { useLocalizedStrings } from '@papi/frontend/react';
 import { CreateProjectModal } from '../../components/CreateProjectModal';
 
-jest.mock('../../components/SelectInterlinearProjectModal', () => ({
-  __esModule: true,
-  /** Minimal re-implementation that avoids importing the real module's coverage into this suite. */
-  isInterlinearProjectSummary(p: unknown): boolean {
-    if (!p || typeof p !== 'object') return false;
-    if (!('id' in p) || typeof p.id !== 'string') return false;
-    if (!('createdAt' in p) || typeof p.createdAt !== 'string') return false;
-    if (!('sourceProjectId' in p) || typeof p.sourceProjectId !== 'string') return false;
-    if (
-      !('analysisLanguages' in p) ||
-      !Array.isArray(p.analysisLanguages) ||
-      !p.analysisLanguages.every((l) => typeof l === 'string')
-    )
-      return false;
-    if ('name' in p && typeof p.name !== 'string') return false;
-    if ('description' in p && typeof p.description !== 'string') return false;
-    return true;
-  },
-}));
-
 const testProjectId = 'test-project-id';
 
 describe('CreateProjectModal', () => {
